Use safer variable names.

+++ projects/run-tests.sh
@@ -2,13 +2,17 @@
 
 set -e
 
-for project in `ls -d */`; do
-    REQUIREMENTS_FILE=$project/requirements.txt
-    echo "Run test for $project."
+for PROJECT in `ls -d */`; do
+    if [ -e "${PROJECT}/.skipci"]; then
+        echo "Skipping ${PROJECT} tests."
+        continue
+    fi
+    REQUIREMENTS_FILE="${PROJECT}/requirements.txt"
+    echo "Run test for ${PROJECT}."
     if [ -e ${REQUIREMENTS_FILE} ]; then
         echo "Installing requirements:"
         cat ${REQUIREMENTS_FILE}
         conda install --yes --use-index-cache --file ${REQUIREMENTS_FILE}
     fi
-    ./flow-test $project -vv --timeout=600 $@
+    ./flow-test ${PROJECT} -vv --timeout=600 $@
 done
